Keep venv in container disk instead of on persistent storage

+++ scripts/pre_start.sh
@@ -27,21 +27,16 @@ sync_apps() {
     fi
 }
 
-fix_venvs() {
-    # Fix the venv to make it work from /workspace
-    echo "Fixing venv..."
-    /fix_venv.sh /kohya_ss/venv /workspace/kohya_ss/venv
-}
-
 link_models() {
-    # Link model
-    ln -s /sd-models/sd_xl_base_1.0.safetensors /workspace/sd_xl_base_1.0.safetensors
+    # Link model if its not already linked
+    if [[ ! -L /workspace/sd_xl_base_1.0.safetensors ]]; then
+        ln -s /sd-models/sd_xl_base_1.0.safetensors /workspace/sd_xl_base_1.0.safetensors
+    fi
 }
 
 if [ "$(printf '%s\n' "$EXISTING_VERSION" "$TEMPLATE_VERSION" | sort -V | head -n 1)" = "$EXISTING_VERSION" ]; then
     if [ "$EXISTING_VERSION" != "$TEMPLATE_VERSION" ]; then
         sync_apps
-        fix_venvs
         link_models
 
         # Configure accelerate
